# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "account-app-sever API",
    "version": "1.0.0",
    "description": "根据当前 account-app-sever 路由实现整理的 OpenAPI 规范。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true }
        },
        "required": ["status", "message", "data"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "请求失败: UnauthorizedError" },
          "errors": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["status", "message", "errors"]
      },
...
      "AccountWriteRequest": {
        "type": "object",
        "properties": {
          "ledger_id": { "type": "integer" },
          "client_id": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "currency": { "type": "string" },
          "opening_balance": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "deleted_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CategoryWriteRequest": {
        "type": "object",
        "properties": {
          "ledger_id": { "type": "integer" },
          "client_id": { "type": "string" },
          "name": { "type": "string" },
          "kind": { "type": "string" },
          "color": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "deleted_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "获取账户列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "integer" },
            "required": false,
            "description": "当用户没有 currentLedgerId 时的兜底账本 ID"
          }
        ],
        "responses": {
          "200": { "description": "获取成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "缺少账本", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "未认证", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "创建账户",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountWriteRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "缺少账本", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "获取单个账户",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "获取成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "账户不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Accounts"],
        "summary": "更新账户",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountWriteRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "账户不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Accounts"],
        "summary": "删除账户（软删除）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "删除成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "账户不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "获取分类列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "integer" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "获取成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "缺少账本", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "创建分类",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryWriteRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "获取单个分类",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "获取成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "分类不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Categories"],
        "summary": "更新分类",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryWriteRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "分类不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "删除分类（软删除）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "删除成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "分类不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}